task 4

+++ supervised_learning/0x01-classification/4-neuron.py
@@ -48,4 +48,4 @@ class Neuron():
         """evaluates neuron predictions"""
         B = self.forward_prop(X)
         cost = self.cost(Y, B)
-        return (np.round(B), cost)
+        return (np.round(B).astype(int), cost)
